+++ conjure-go-client/httpclient/response_error_decoder_middleware_test.go
@@ -201,89 +201,6 @@ func TestErrorDecoderMiddlewares(t *testing.T) {
 	}
 }
 
-func TestGetStatusCodeFromError(t *testing.T) {
-	for _, tc := range []struct {
-		name                  string
-		err                   error
-		expectStatusCodeExist bool
-		expectStatusCode      int
-	}{
-		{
-			name:                  "no status code",
-			err:                   werror.Error("200"),
-			expectStatusCodeExist: false,
-			expectStatusCode:      0,
-		},
-		{
-			name: "status code 200",
-			err: werror.Error("200",
-				werror.SafeParam("statusCode", 200)),
-			expectStatusCodeExist: true,
-			expectStatusCode:      200,
-		},
-		{
-			name: "status code non int",
-			err: werror.Error("200",
-				werror.SafeParam("statusCode", "200")),
-			expectStatusCodeExist: false,
-			expectStatusCode:      0,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			code, exist := httpclient.StatusCodeFromError(tc.err)
-			assert.Equal(t, tc.expectStatusCodeExist, exist)
-			assert.Equal(t, tc.expectStatusCode, code)
-		})
-	}
-}
-
-func TestGetLocationFromError(t *testing.T) {
-	for _, tc := range []struct {
-		name                string
-		err                 error
-		expectLocationExist bool
-		expectLocation      string
-	}{
-		{
-			name: "200 no location",
-			err: werror.Error("200",
-				werror.SafeParam("statusCode", 200)),
-			expectLocationExist: false,
-			expectLocation:      "",
-		},
-		{
-			name: "307 with location",
-			err: werror.Error("307",
-				werror.SafeParam("statusCode", 307),
-				werror.UnsafeParam("location", "https://google.com")),
-			expectLocationExist: true,
-			expectLocation:      "https://google.com",
-		},
-		{
-			name: "307 without location",
-			err: werror.Error("307",
-				werror.SafeParam("statusCode", 307),
-				werror.UnsafeParam("location", "")),
-			expectLocationExist: true,
-			expectLocation:      "",
-		},
-		{
-			name: "307 with non string location",
-			err: werror.Error("307",
-				werror.SafeParam("statusCode", 307),
-				werror.UnsafeParam("location", 12345)),
-			expectLocationExist: false,
-			expectLocation:      "",
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			location, exist := httpclient.LocationFromError(tc.err)
-			assert.Equal(t, tc.expectLocationExist, exist)
-			assert.Equal(t, tc.expectLocation, location)
-		})
-	}
-}
-
 type fooErrorDecoder struct{}
 
 func (d fooErrorDecoder) Handles(resp *http.Response) bool {
